# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/me": {
      "get": {
        "operationId": "AuthController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/invoices/draft": {
      "post": {
        "operationId": "InvoicesController_draft",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInvoiceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Draft created"
          },
          "400": {
            "description": "Validation error"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create invoice draft and compute totals",
        "tags": [
          "Invoices"
        ]
      }
    },
    "/invoices/{id}/confirm": {
      "post": {
        "operationId": "InvoicesController_confirm",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-idempotency-key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmInvoiceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Invoice confirmed (idempotent)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Confirm draft invoice with DGI checks",
        "tags": [
          "Invoices"
        ]
      }
    },
    "/invoices/{id}": {
      "get": {
        "operationId": "InvoicesController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get invoice by id"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Invoices"
        ]
      }
    },
    "/invoices": {
      "get": {
        "operationId": "InvoicesController_list",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List invoices"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Invoices"
        ]
      }
    },
    "/invoices/{id}/normalized": {
      "get": {
        "operationId": "InvoicesController_normalized",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get normalized payload with sha256"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Invoices"
        ]
      }
    },
    "/invoices/{id}/pdf": {
      "get": {
        "operationId": "InvoicesController_pdfBinary",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "PDF binary"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Generate invoice PDF",
        "tags": [
          "Invoices"
        ]
      }
    },
...
    "/me/profile": {
      "get": {
        "operationId": "ProfileController_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Profile"
        ]
      },
      "put": {
        "operationId": "ProfileController_updateProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Profile"
        ]
      }
    },
    "/me/password": {
      "patch": {
        "operationId": "ProfileController_changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Profile"
        ]
      }
    },
    "/me/roles": {
      "get": {
        "operationId": "ProfileController_myRoles",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Profile"
        ]
      }
    },
    "/me/permissions": {
      "get": {
        "operationId": "ProfileController_myPermissions",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Profile"
        ]
      }
    },
...
  },
  "info": {
    "title": "SFE API",
    "description": "OpenAPI autogénéré depuis le code",
    "version": "v1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "X-Tenant-Id": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Tenant-Id",
        "description": "Tenant context header"
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string",
            "minLength": 8
          },
          "organizationName": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string",
            "minLength": 8
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
...
      "InvoiceClientDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "PP",
              "PM",
              "PC",
              "PL",
              "AO"
            ]
          },
          "denomination": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "nif": {
            "type": "string"
          },
          "refExo": {
            "type": "string"
          }
        },
        "required": [
          "type"
        ]
      },
      "InvoiceLineDto": {
        "type": "object",
        "properties": {
          "itemId": {
            "type": "string"
          },
          "kind": {
            "type": "string",
            "enum": [
              "BIE",
              "SER",
              "TAX"
            ]
          },
          "group": {
            "type": "string",
            "enum": [
              "A",
              "B",
              "C",
              "D",
              "E",
              "F",
              "G",
              "H",
              "I",
              "J",
              "K",
              "L",
              "M",
              "N",
              "O",
              "P"
            ]
          },
          "label": {
            "type": "string"
          },
          "qty": {
            "type": "string",
            "description": "quantity (string), scale 3"
          },
          "unitPrice": {
            "type": "string",
            "description": "unit price (string), scale 2"
          }
        },
        "required": [
          "kind",
          "group",
          "qty",
          "unitPrice"
        ]
      },
      "InvoiceAvoirDto": {
        "type": "object",
        "properties": {
          "nature": {
            "type": "string",
            "enum": [
              "COR",
              "RAN",
              "RAM",
              "RRR"
            ]
          },
          "originInvoiceRef": {
            "type": "string"
          }
        }
      },
      "CreateInvoiceDto": {
        "type": "object",
        "properties": {
          "modePrix": {
            "type": "string",
            "enum": [
              "HT",
              "TTC"
            ]
          },
          "type": {
            "type": "string",
            "enum": [
              "FV",
              "FT",
              "FA",
              "EV",
              "ET",
              "EA"
            ]
          },
          "client": {
            "$ref": "#/components/schemas/InvoiceClientDto"
          },
          "lines": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InvoiceLineDto"
            }
          },
          "avoir": {
            "$ref": "#/components/schemas/InvoiceAvoirDto"
          }
        },
        "required": [
          "modePrix",
          "type",
          "client",
          "lines"
        ]
      },
      "EquivalentCurrencyDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "Currency code (quote) like USD",
            "example": "USD"
          }
        },
        "required": [
          "code"
        ]
      },
      "ConfirmInvoiceDto": {
        "type": "object",
        "properties": {
          "equivalentCurrency": {
            "$ref": "#/components/schemas/EquivalentCurrencyDto"
          }
        }
      },
...
      "UpdateProfileDto": {
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string",
            "minLength": 2,
            "maxLength": 80
          },
          "phone": {
            "type": "string",
            "minLength": 5,
            "maxLength": 40
          },
          "avatarUrl": {
            "type": "string",
            "format": "uri"
          },
          "locale": {
            "type": "string",
            "minLength": 2,
            "maxLength": 10
          },
          "timezone": {
            "type": "string",
            "minLength": 2,
            "maxLength": 40
          }
        }
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {
          "currentPassword": {
            "type": "string"
          },
          "newPassword": {
            "type": "string"
          }
        },
        "required": [
          "currentPassword",
          "newPassword"
        ]
      },
...
    }
  },
  "security": [
    {
      "X-Tenant-Id": []
    }
  ]
}